Rename dataforms-layouts/ to dataform-layout/ for consistency with any other folder in the directory
