+++ packages/dataviews/src/components/dataform/index.tsx
@@ -9,7 +9,7 @@ import { useMemo } from '@wordpress/element';
 import type { DataFormProps } from '../../types';
 import { DataFormProvider } from '../dataform-context';
 import { normalizeFields } from '../../normalize-fields';
-import { DataFormLayout } from '../../dataforms-layouts/data-form-layout';
+import { DataFormLayout } from '../../dataform-layouts/data-form-layout';
 
 export default function DataForm< Item >( {
 	data,
