+++ packages/dataviews/src/dataform-layouts/get-summary-fields.ts
