+++ packages/dataviews/src/style.scss
@@ -14,7 +14,7 @@
 @import "./dataviews-layouts/picker-grid/style.scss";
 
 @import "./dataform-controls/style.scss";
-@import "./dataforms-layouts/panel/style.scss";
-@import "./dataforms-layouts/regular/style.scss";
-@import "./dataforms-layouts/card/style.scss";
-@import "./dataforms-layouts/row/style.scss";
+@import "./dataform-layouts/panel/style.scss";
+@import "./dataform-layouts/regular/style.scss";
+@import "./dataform-layouts/card/style.scss";
+@import "./dataform-layouts/row/style.scss";
